requirements.txt
jdatetime==5.2.0
openai==1.97.1
openpyxl==3.1.5
pandas==2.3.1
python-dotenv==1.1.1
pymongo==4.13.2
